+++ CHANGES.md
@@ -7,7 +7,8 @@ See [compatibility docs](docs/compatibility.md) for details about backward compa
 ## CHANGE LOG ##
 
 * @dev
-   * Separate `ExceptionHandler::TYPE_HTTP_NOT_FOUND_KEY` and all related stuff, incl. localization key `http_not_found`,
+   * Reworked `ExceptionHandlerHelper` configuration. Now, you will be able to easily configure every
+     HttpException for each HTTP status code you want. Separate `ExceptionHandler::TYPE_HTTP_NOT_FOUND_KEY` and all related stuff, incl. localization key `http_not_found`,
      configuration is now replace with more flexible generic code that provides error messages for all supported
      HTTP codes from in range `400-599`.